Add dotenv and set vim executable and return function
This adds the 'dotenv' gem to require setting a vim executable path.
Also, when using MacVim, MacVim steals focus during tests and then does
not return it.

This allows someone to set a "return command" that will be run from vim
to go back to the calling application. For example:
`open /Applications/Alacritty.app`.