Add topic builder, topic objects, and folds
This is a lot, it's added from a first draft of this project.

It adds objects in a `lib` folder, right now it has metadata and a
topic.

A topic gets built with a metadata, title, description, title, level and
line number.

For now, this is mostly just used to set some fold text, but the hope is
to maybe have some more interactive pages that can be built from the
data structures that are built.

It also adds a builder which builds stuff line by line.