Update autoline logic
Previously, tada was not taking advantage of the 'comments=' command
which can be used to auto insert text as a comment. This adds that
logic, which greatly simplifies how to handle autolines. (Automatically
added '-', '|', and '- [ ]' prefixes.)

I had hoped that we could use the comment string for todo boxes as well,
but given that for a completed to do, we would still want to insert a
blank todo, it doesn't work. The commant string does not take patterns.
We could build a string to recognize all the symbols, but they would
then only insert the came symbol.

In theory, we could use it to handle some of the logic, but then we
would need to replace the symbol with the blank one, and I'm not sure
that is any simpler.