Bugfix: Don't infinitely look, set local methods
1st, we are going to set all settings at `setlocal` instead of `set` as
it should be specific to the tada file.

Additionally, fix a bad bug where if we set custom statuses and
symbols, and our file has a todo line that doesn't match those, it would
cause vim to look for forever because we weren't stopping before the end
of the file.

Realistically, the builder should just iterate over each line, checking
what type it is an then returning, instead of assuming the order, but
that was a choice that works fow now.