Remove leading '-' when a '|' is typed on an empty line
In the next commit of "doing too much", if you are on an empty line with
a leading "- " and type a "|" this will delete the leading dash and
space. The idea being if you want to type metadata, you don't need to go
backwards to delete the leading list item syntax.

In theory, this makes is _really_ difficult to type a pipe as a leading
character, but I think the trade off is worth it.