Adjust fold calculators to use regexes instead of syntax groups
When testing folding on a 30k line file, it was noticeably slower. Some
of that is to be expected for size. When I added simple folding with no
lines or fancy calculations (just every 30 lines would start a fold) it
still took some time on a large file.

However, there was a noticeable difference when using the syntax group
of a topic vs just testing for a number of spaces and a colon. Likewise,
we calculate the fold level to compare against the g:tada_fold_level
using syntax groups as well. Now we just do a mod 2 of the indent. And
it seems to be significantly faster.

Unfortunately all of these tests are on an M1 mac, so I'm curious how it
is on a non M1 if it's even in the realm of tolerable.