Add json methods for all data points
Not sure if this will be useful, but it adds ToJson methods for the 3
data objects. This could be used to construct a UI or something else
like that.

I was trying to make this work in markdown format, but I prefer the
nested todo list. Markdown would require headers and it changes the
structure, so producing data means we could maybe produce a file that
could be html or something like that which could be viewed.