Pause to deal with vimrunner bug
For some reason, between these two tests, vim starts too fast and the GA
in the second test is run in insert mode (despite being run with
vim.normal). Adding a sleep or an esc or a normal "" all would fix this,
but for now we are just adding a 'normal ""' before all specs.