Update fold logic
The fold logic was previously unnecessarily complicated.

We can set nested fold levels and just use the regular fold commands to
achieve what we are looking for. Enter will close a fold level, or a few
keymaps can replace the normal fold ones if the <C-T> prefix is easier
to remember.

Additionally, this automatically folds archived todos.