Refactor box logic, fix bugs
There were a few bugs with the <C-B> logic depending what was in the
line.

Additionally, I just learned about `<Cmd>` for keymaps which will
simplify a lot of things, going to move over to that. It will solve the
issue of virtual edits when you're at the end of the line but a column
does not actually exist.